refactored course management page components
